<==== Mercurial data ====>
Branch: release_2.3

+++ appserver2/src/connection_factory.cpp
@@ -257,7 +257,7 @@ namespace ec2
         registerGetFuncHandler<std::nullptr_t, ApiLicenseDataList>( restProcessorPool, ApiCommand::getLicenses );
 
         registerGetFuncHandler<std::nullptr_t, ApiDatabaseDumpData>( restProcessorPool, ApiCommand::dumpDatabase );
-        registerGetFuncHandler<ApiStoredFilePath, size_t>( restProcessorPool, ApiCommand::dumpDatabaseToFile );
+        registerGetFuncHandler<ApiStoredFilePath, qint64>( restProcessorPool, ApiCommand::dumpDatabaseToFile );
 
         //AbstractECConnectionFactory
         registerFunctorHandler<ApiLoginData, QnConnectionInfo>( restProcessorPool, ApiCommand::connect,
